REFACTOR : 미사용 코드 제거 및 구조 정리
